Implement "dumb" minimax for AI moves
Implements minimax for AI, but some strange behaviors persists and
therefore still needs optimizations in minimax logic and weights.

Issue: AI gets trapped in a loop when the opponent makes moves on the
side of the board, and tends to make moves in that area even after
the opponent has given up on that region.